Option to show slide numbers
Slide numbers can be enabled by either using the "-N"/"--slide-numbers"
command line switch or by putting ":slide-numbers: true" in the
presentation preamble. The are shown in the bottom right corner of the
presentation by default and can be customised using CSS. In the
presenter console (default template), the current slide number is shown
next to the status ("Moving"/"Ready").